[rename]구분을 위해 SignupPage 폴더로 이동
